LifeSimultor.jl

Documentation to be added.

API

LifeSimulator.CashFlow

LifeSimulator.LifelibBasiclife

LifeSimulator.LifelibSavings

LifeSimulator.Model

LifeSimulator.Policy

LifeSimulator.PolicySet

LifeSimulator.Simulation

LifeSimulator.SimulationEvents

LifeSimulator.UniversalLifeModel

LifeSimulator.next!

LifeSimulator.next!

LifeSimulator.policies_from_csv

LifeSimulator.CashFlow — Type

Present value of future cashflows as computed by a Simulation over a UniversalLifeModel.

The present value of future cashflows is estimated from the future cashflows using a model-provided discount rate.

LifeSimulator.LifelibBasiclife — Type

Term life insurance model replicating the functionality of lifelib's basiclife module.

LifeSimulator.LifelibSavings — Type

Universal life model reimplemented from lifelib's savings library.

LifeSimulator.Model — Type

Model defining the evolution of policies with regards to lapses, cashflows and, for UniversalLifeModels, account values for policy holders.

Models are to be evaluated using a Simulation.

See also: Policy

LifeSimulator.Policy — Type

Policy held with a corresponding account value.

LifeSimulator.PolicySet — Type

Aggregation of a specific type of policy among many policy holders.

The account for universal life models is therefore also unique among all policy holders, amounting to a total of set.policy.account_value * policy_count(set)

LifeSimulator.Simulation — Type

Simulation parametrized by a particular Model.

The simulation time starts at the current date by default. The simulation is carried out every month, producing events ([SimulationEvents]) corresponding to what happened between two timesteps, i.e. from one month to the other. The simulation is nonetheless stateful, meaning that such events may only be produced once; the next evaluation will return the events for the timestep after that.

See also: next!

LifeSimulator.SimulationEvents — Type

Events that happen as part of a simulation timestep for a Model.

These events are meant to be processed by the user in order to generate quantities of interest that do not involve simulation-related state. This is, for example, how the CashFlow quantities are computed.

LifeSimulator.UniversalLifeModel — Type

Universal life model.

The defining property of a universal life model is that the contract between an insurance company and a policy holder involves a client-managed bank account, where the policy holder is responsible for keeping the bank account appropriately filled.

Lapses do not occur in absence of a payment, but rather when the bank account runs out of money while fees or premiums must be paid. It is, for example, for the policy holder to put a lot of money in the bank account and forget about it for a while, as opposed to a regular term life insurance model which requires frequent payments without such buffer.

LifeSimulator.next! — Method

Perform a simulation timestep over the LifelibBasiclife model, returning a SimulationEvents.

First, the policies which reached their term are removed, yielding claims and account changes.

Second, the policies which start from the current month are added, yielding expenses (costs for the insurance company).

Then, at the middle of the month, deaths and lapses occur. Finally, the simulation time is incremented.

A callback may be run just before the deaths and lapses occur, as the original basiclife model considers lapses and deaths to be part of the next iteration (i.e., deaths and lapses occur prior to the next step, and not in the current step).

LifeSimulator.next! — Method

Perform a simulation timestep over the LifelibSavings model, returning a SimulationEvents.

First, the policies which reached their term are removed, yielding claims and account changes.

Second, the policies which start from the current month are added, yielding expenses (costs for the insurance company).

Third, all account values are updated, with:

A premium amount put into the bank account (minus fees, the load premium rate).

Maintenance fees withdrawn from the back account.

Insurance costs withdrawn from the back account.

Investments realized during the previous month.

Then, at the middle of the month, deaths and lapses occur. Finally, the simulation time is incremented.

LifeSimulator.policies_from_csv — Function

Import policies from CSV files compatible with the lifelib model.